+++ src/test/scala/com/rayrobdod/json/builder/BuildableBuilderTest.scala
@@ -50,7 +50,7 @@ class BuildableBuilderTest extends FunSpec {
 			assertResult(Right(new Person(name, 0))){
 				new BuildableBuilder(new Person("", 0))
 						.addDef("name", new KeyDef[String, String, Person]{def apply[I](s:Person, i:I, p:Parser[String, String, I]) = {Right(s.copy(name = p.parsePrimitive(i).fold({x => ""}, {x => x})))}})
-						.apply(new Person("", 0), "name", name, new IdentityParser)
+						.apply(new Person("", 0), "name", name, new IdentityParser[String])
 			}
 		}
 		it ("Acts upon provided keydef (2)") {
@@ -58,7 +58,7 @@ class BuildableBuilderTest extends FunSpec {
 			assertResult(Right(new Person("", age))){
 				new BuildableBuilder(new Person("", 0))
 						.addDef("age", new KeyDef[String, Int, Person]{def apply[I](s:Person, i:I, p:Parser[String, Int, I]) = {Right(s.copy(age = p.parsePrimitive(i).fold({x => 0}, {x => x})))}})
-						.apply(new Person("", 0), "age", age, new IdentityParser)
+						.apply(new Person("", 0), "age", age, new IdentityParser[Int])
 			}
 		}
 		it ("Acts upon provided keydef (partitionedPrimitive)") {
@@ -66,21 +66,21 @@ class BuildableBuilderTest extends FunSpec {
 			assertResult(Right(new Person("", age))){
 				new BuildableBuilder(new Person("", 0))
 						.addDef("age", BuildableBuilder.partitionedPrimitiveKeyDef[String, Int, Person, Int]({case x:Int => Right(x)}, {(f,i) => f.copy(age = i)}))
-						.apply(new Person("", 0), "age", age, new IdentityParser)
+						.apply(new Person("", 0), "age", age, new IdentityParser[Int])
 			}
 		}
 		it ("Throws excpetion on unknown key") {
 			val age = "9001"
 			assertResult(Left("BuildableBuilder has no KeyDef for given key", 0)){
 				new BuildableBuilder[String, String, Person](new Person("", 0))
-						.apply(new Person("", 0), "asdfjkl;", "hello", new IdentityParser)
+						.apply(new Person("", 0), "asdfjkl;", "hello", new IdentityParser[String])
 			}
 		}
 		it ("ignores unknown key after call to ignoreUnknownKeys") {
 			val age = "9001"
 			assertResult(Right(new Person("", 0))){
 				new BuildableBuilder[String, String, Person](new Person("", 0)).ignoreUnknownKeys
-						.apply(new Person("", 0), "asdfjkl;", "hello", new IdentityParser)
+						.apply(new Person("", 0), "asdfjkl;", "hello", new IdentityParser[String])
 			}
 		}
 	}
@@ -136,18 +136,18 @@ class BuildableBuilderTest extends FunSpec {
 		it ("partitionedPrimitiveKeyDef (isDefinedAt)") {
 			val builder = partitionedPrimitiveKeyDef[Any, String, Option[Int], Int]({case "abc" => Right(3)}, {(a,b) => Some(b)})
 			assertResult(Right(Some(3))){
-				builder.apply(None, "abc", new IdentityParser)
+				builder.apply(None, "abc", new IdentityParser[String])
 			}
 		}
 		it ("partitionedPrimitiveKeyDef (not isDefinedAt)") {
 			val builder = partitionedPrimitiveKeyDef[Any, String, Option[Int], Int]({case "abc" => Right(3)}, {(a,b) => Some(b)})
 			assertResult(Left("Unexpected value: asdf", 0)){
-				builder.apply(None, "asdf", new IdentityParser)
+				builder.apply(None, "asdf", new IdentityParser[String])
 			}
 		}
 		it ("partitionedKeyDef (isDefinedAt)") {
 			val builder = partitionedKeyDef[Int, Int, String, Seq[Int], Int](
-				new PrimitiveSeqBuilder[Int, Int],
+				new PrimitiveSeqBuilder[Int],
 				{case ParserRetVal.Complex(x) => Right(x.sum)},
 				{(a,b) => a + " " + b}
 			)
@@ -157,7 +157,7 @@ class BuildableBuilderTest extends FunSpec {
 		}
 		it ("partitionedKeyDef (not isDefinedAt)") {
 			val builder = partitionedKeyDef[Int, Int, String, Seq[Int], Int](
-				new PrimitiveSeqBuilder[Int, Int],
+				new PrimitiveSeqBuilder[Int],
 				{case ParserRetVal.Primitive(x) => Right(x)},
 				{(a,b) => a + " " + b}
 			)
@@ -177,7 +177,7 @@ class BuildableBuilderTest extends FunSpec {
 		}
 		it ("partitionedKeyDef (isDefinedAt; left)") {
 			val builder = partitionedKeyDef[Int, Int, String, Seq[Int], Int](
-				new PrimitiveSeqBuilder[Int, Int],
+				new PrimitiveSeqBuilder[Int],
 				{case ParserRetVal.Complex(x) => Left("Error", 0)},
 				{(a,b) => a + " " + b}
 			)
@@ -187,7 +187,7 @@ class BuildableBuilderTest extends FunSpec {
 		}
 		it ("partitionedComplexKeyDef (Complex)") {
 			val builder = partitionedComplexKeyDef[Int, Int, String, Seq[Int]](
-				new PrimitiveSeqBuilder[Int, Int],
+				new PrimitiveSeqBuilder[Int],
 				{(a,b) => Right(a + " " + b)}
 			)
 			assertResult(Right("z Vector(1, 2, 3)")){
@@ -196,11 +196,11 @@ class BuildableBuilderTest extends FunSpec {
 		}
 		it ("partitionedComplexKeyDef (Primitive)") {
 			val builder = partitionedComplexKeyDef[Int, Int, String, Seq[Int]](
-				new PrimitiveSeqBuilder[Int, Int],
+				new PrimitiveSeqBuilder[Int],
 				{(a,b) => Right(a + " " + b)}
 			)
 			assertResult(Left("Unexpected value: 5", 0)){
-				builder.apply("z", 5, new IdentityParser[Int, Int])
+				builder.apply("z", 5, new IdentityParser[Int])
 			}
 		}
 		it ("partitionedComplexKeyDef (Throw)") {