+++ src/test/scala/com/rayrobdod/json/builder/CborObjectBuilderTest.scala
@@ -49,122 +49,122 @@ class CborObjectBuilderTest extends FunSpec {
 		}
 		it ("Appends null") {
 			assertResult(Right(hexSeq"A160f6")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValueNull, new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValueNull, new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends true") {
 			assertResult(Right(hexSeq"A160f5")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(true), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(true), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends false") {
 			assertResult(Right(hexSeq"A160f4")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(false), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(false), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends 0 as one byte") {
 			assertResult(Right(hexSeq"A16000")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(0), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(0), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends small positive integer") {
 			assertResult(Right(hexSeq"A16006")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(6), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(6), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends 23 as one byte") {
 			assertResult(Right(hexSeq"A16017")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(23), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(23), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends 24 as two bytes") {
 			assertResult(Right(hexSeq"A1601818")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(24), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(24), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends 255 as two bytes") {
 			assertResult(Right(hexSeq"A16018FF")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(255), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(255), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends 256 as three bytes") {
 			assertResult(Right(hexSeq"A160190100")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(256), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(256), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends -1 as one byte") {
 			assertResult(Right(hexSeq"A16020")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(-1), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(-1), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends -100 as two bytes") {
 			assertResult(Right(hexSeq"A1603863")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(-100), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(-100), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends Float") {
 			assertResult(Right(hexSeq"A160 FA3FC00000")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(1.5f), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(1.5f), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends Double") {
 			assertResult(Right(hexSeq"A160 fbc010666666666666")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(-4.1D), new IdentityParser[CborValue, CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(-4.1D), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends empty string") {
 			assertResult(Right(hexSeq"A16060")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(""), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(""), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends \"abc\"") {
 			assertResult(Right(hexSeq"A16063616263")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue("abc"), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue("abc"), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends string containing multibyte char") {
 			assertResult(Right(hexSeq"A16063e6b0b4")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue("\u6c34"), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue("\u6c34"), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends an empty byte array") {
 			assertResult(Right(hexSeq"A16040")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(new Array[Byte](0)), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(new Array[Byte](0)), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends a non-empty byte array") {
 			assertResult(Right(hexSeq"A16043010203")){
-				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(hexArray"010203"), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"A0", CborValue(""), CborValue(hexArray"010203"), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends a second value") {
 			assertResult(Right(hexSeq"""A26161616261636164""")){
-				new CborObjectBuilder().apply(hexSeq"A161616162", CborValue("c"), CborValue("d"), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"A161616162", CborValue("c"), CborValue("d"), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends a 24th value") {
 			assertResult(Right(hexSeq"""B818 B7A8B7A8B7A8 6060""")){
-				new CborObjectBuilder().apply(hexSeq"""B7 B7A8B7A8B7A8 """, CborValue(""), CborValue(""), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"""B7 B7A8B7A8B7A8 """, CborValue(""), CborValue(""), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends a 101th value") {
 			assertResult(Right(hexSeq"""B903e9 B7A8B7A8B7A8 6060""")){
-				new CborObjectBuilder().apply(hexSeq"""B903e8 B7A8B7A8B7A8""", CborValue(""), CborValue(""), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"""B903e8 B7A8B7A8B7A8""", CborValue(""), CborValue(""), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends a 256th value") {
 			assertResult(Right(hexSeq"""B90100 B7A8B7A8B7A8 6060""")){
-				new CborObjectBuilder().apply(hexSeq"""B8FF B7A8B7A8B7A8""", CborValue(""), CborValue(""), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"""B8FF B7A8B7A8B7A8""", CborValue(""), CborValue(""), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends a 65536th value") {
 			assertResult(Right(hexSeq"""BA00010000 B7A8B7A8B7A8 6060""")){
-				new CborObjectBuilder().apply(hexSeq"""B9FFFF B7A8B7A8B7A8""", CborValue(""), CborValue(""), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"""B9FFFF B7A8B7A8B7A8""", CborValue(""), CborValue(""), new IdentityParser[CborValue])
 			}
 		}
 		it ("Appends a 1000001th value") {
 			assertResult(Right(hexSeq"""Ba000f4241 B7A8B7A8B7A8 6060""")){
-				new CborObjectBuilder().apply(hexSeq"""Ba000f4240 B7A8B7A8B7A8""", CborValue(""), CborValue(""), new IdentityParser[CborValue,CborValue])
+				new CborObjectBuilder().apply(hexSeq"""Ba000f4240 B7A8B7A8B7A8""", CborValue(""), CborValue(""), new IdentityParser[CborValue])
 			}
 		}