+++ src/test/scala/com/rayrobdod/json/parser/CsvParserTest_Happy.scala
@@ -77,12 +77,12 @@ final class CsvParserTest_Happy extends FunSpec {
 			
 			it (name) {
 				val source = source2
-				val result = new CsvParser(charMeans).parse(new SeqBuilder(new PrimitiveSeqBuilder[Int, String]), source).fold({x => x},{x => x},{(a,b) => a})
+				val result = new CsvParser(charMeans).parse(new SeqBuilder(new PrimitiveSeqBuilder[String]), source).fold({x => x},{x => x},{(a,b) => a})
 				assertResult(expected){result}
 			}
 			it (name + " (reader)") {
 				val source = new java.io.StringReader(source2)
-				val result = new CsvParser(charMeans).parse(new SeqBuilder(new PrimitiveSeqBuilder[Int, String]), source).fold({x => x},{x => x},{(a,b) => a})
+				val result = new CsvParser(charMeans).parse(new SeqBuilder(new PrimitiveSeqBuilder[String]), source).fold({x => x},{x => x},{(a,b) => a})
 				assertResult(expected){result}
 			}
 		}