+++ src/test/scala/com/rayrobdod/json/parser/JsonParserTest_Unhappy.scala
@@ -221,25 +221,25 @@ class JsonParserTest_Unhappy extends FunSpec {
 		it ("""provides a correct index in a nested value (array, array)""") {
 			val source = """[[,]]"""
 			assertFailureParse("",2){
-				new JsonParser().parse(new SeqBuilder(new PrimitiveSeqBuilder[StringOrInt, JsonValue]), source)
+				new JsonParser().parse(new SeqBuilder(new PrimitiveSeqBuilder[JsonValue]), source)
 			}
 		}
 		it ("""provides a correct index in a nested value (array, object)""") {
 			val source = """[{},{,}]"""
 			assertFailureParse("",5){
-				new JsonParser().parse(new SeqBuilder(new PrimitiveSeqBuilder[StringOrInt, JsonValue]), source)
+				new JsonParser().parse(new SeqBuilder(new PrimitiveSeqBuilder[JsonValue]), source)
 			}
 		}
 		it ("""provides a correct index in a nested value (object, array)""") {
 			val source = """{"":[,]}"""
 			assertFailureParse("",5){
-				new JsonParser().parse(new SeqBuilder(new PrimitiveSeqBuilder[StringOrInt, JsonValue]), source)
+				new JsonParser().parse(new SeqBuilder(new PrimitiveSeqBuilder[JsonValue]), source)
 			}
 		}
 		it ("""provides a correct index in a nested value (object, object)""") {
 			val source = """{"" : {,}}"""
 			assertFailureParse("",7){
-				new JsonParser().parse(new SeqBuilder(new PrimitiveSeqBuilder[StringOrInt, JsonValue]), source)
+				new JsonParser().parse(new SeqBuilder(new PrimitiveSeqBuilder[JsonValue]), source)
 			}
 		}
 		it ("""Throwbuilder (array of string)""") {