+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 3.0-RC2
-* Documentation, mostly
 * Isolate CborValueByteStr backing value from inputs or outputs
+* Figure out how to make Builder and Parser co- and contravariant
+* Add projections to ParserRetVal cases
 
 ## 3.0-RC1
 Typesafety